+++ Samples~/Scripts/CleanSlider.cs
@@ -3,12 +3,12 @@ using UnityEngine.UI;
 
 namespace SaintsBuild.Samples
 {
-    public class CleanSlider: MonoBehaviour, IPostProcessScene
+    public class CleanSlider: MonoBehaviour, IPostProcess
     {
         public Slider slider;
 
 #if UNITY_EDITOR
-        public void EditorOnPostProcessScene(bool isBuilding)
+        public void EditorOnPostProcess(PostProcessInfo postProcessInfo)
         {
             slider.value = 0;
         }